pyproject.toml
[tool.poetry]
name = "django-demo"
version = "0.1.0"
description = ""
authors = ["Roman Hahn <roman@frolov.me>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.11.3"
django = "^4.2.3"
djangorestframework-simplejwt = "^5.2.2"
django-behaviors = "^0.5.1"
django-environ = "^0.10.0"
django-filter = "^23.2"
django-healthchecks = "^1.5.0"
django-split-settings = "^1.2.0"
django-storages = "^1.13.2"
djangorestframework = "^3.14.0"
drf-jwt = "^1.19.2"
drf-yasg = "^1.21.6"
psycopg2 = "^2.9.6"
django-map-widgets = "^0.4.1"
django-admin-list-filter-dropdown = "^1.0.3"
pillow = "^10.0.0"
django-admin-inline-paginator = "^0.4.0"
django-role-permissions = "^3.2.0"
gunicorn = "^20.1.0"
uvicorn = "^0.23.0"
coreapi = "^2.3.3"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
flake8-bugbear = "^23.7.10"
flake8-commas = "^2.1.0"
flake8-django = "^1.3"
flake8-eradicate = "^1.5.0"
flake8-isort = "^6.0.0"
isort = "^5.12.0"
flake8-print = "^5.0.0"
flake8-printf-formatting = "^1.1.2"
flake8-quotes = "^3.3.2"
flake8-variables-names = "^0.0.6"
ipython = "^8.14.0"
psycopg2-binary = "^2.9.6"
django-stubs = "^4.2.3"
djangorestframework-stubs = "^3.14.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
faker = "^19.1.0"
pytest = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.autopep8]
max_line_length = 120
in-place = true
recursive = true
aggressive = 3


[tool.mypy]
plugins = ['mypy_django_plugin.main']
incremental = true
strict_optional = true
strict_equality = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true

[tool.django-stubs]
django_settings_module = "app.settings"


[tool.isort]
skip = 'venv'
indent = 4
line_length = 160
multi_line_output = 5
combine_as_imports = true
include_trailing_comma = true


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'app.settings'
python_files = 'test*.py'
addopts = ['--strict-markers',]
norecursedirs = [
    'media', 'static', 'migrations', '__pycache__', '.git', 'venv', '.mypy_cache',]
filterwarnings = [
    'ignore::django.utils.deprecation.RemovedInDjango51Warning',
    'ignore::django.utils.deprecation.RemovedInDjango50Warning',]